+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 - added clarification for generating cache with an Alyx database in documentation
 - debug log of exceptions upon HTTP and connection errors during loading of cache
 - added htsv file support in alf.io.load_file_content
+- fix'd dimension check for DataFrame attributes in ALF objects
+- dimension warning logged when dimensions don't match after appending to ALFBunch
+- created_time now updated correctly in cache meta
 
 ## [1.11.0]